Unit test and mocks
